Support polkit session agent running outside user session
commit a68f5dfd7662767b7b9822090b70bc5bd145c50c made
session applications that are running from a user bus
work with polkitd, by falling back to using the currently
active session.

This commit is similar, but for the polkit agent.  It allows,
a polkit agent to be run from a systemd --user service
that's not running directly in the users session.

https://bugs.freedesktop.org/show_bug.cgi?id=96977